feat(inngest): client-free cron manifest leaf + manual-trigger allowlist
Extract EXPECTED_CRON_FUNCTIONS + manualTriggerEventFor into a client-free
cron-manifest.ts leaf so the trigger-cron route + allowlist can import them
without transitively loading the Inngest client (which throws on missing
INNGEST_SIGNING_KEY at module load). The watchdog re-exports both symbols so
its existing importers stay green.

Allowlist (lib/inngest/manual-trigger-allowlist.ts) derives MANUAL_TRIGGER_EVENTS
from the manifest — no parallel hardcoded list to drift.
